Make Transport and SyncTransport Closeable
- I prefer Closeable to AutoCloseable since it throws proper exception (IOException)
- SyncTransport renamed to RawSyncTransport, except one constructor
